refactor: change productId data type (int -> Long)

+++ src/main/java/com/danny/springe_com/model/dto/OrderItemRequest.java
@@ -1,6 +1,6 @@
 package com.danny.springe_com.model.dto;
 
 public record OrderItemRequest(
-        int productId,
+        Long productId,
         int quantity
 ) { }
